refactor: 강의 기간 (SessionPeriod) 엔티티 분리
